managed_context_ops: carve apply.rs (pure move)
Applying an external context rewind and driving what follows: failure
emission, chained resume turns, side/steer follow-up turns, and child
turn-complete events; the RecordingExternalAgent test double moves as
a unit (struct + trait impl). Drops the unused crate-glob scaffold
from the anchors tests.
